separate server from client
