+++ server.js
@@ -1,9 +1,14 @@
-const db = require('../database/index.js');
 const express = require('express');
-const app = express();
+const cors = require('cors');
+// const path = require('path');
+const db = require('./database.js');
 
-app.use(express.static(__dirname + '/../src'));
-app.use(express.json());
+const app = express();
+app.use(cors());
+
+app.set("port", process.env.PORT || 3001);
+// app.use(express.static(__dirname + '/../src'));
+// app.use(express.json());
 
 async function postImage(url, time, tag, index) {
   const image = db`
@@ -25,6 +30,7 @@ async function getImages(tag) {
 
 // post new foaming image to DB
 app.post('/foaming', function(req, res) {
+  console.log('receiving foaming POST request')
   const {url, time, index, tag} = req.body;
   postImage(url, time, tag, index)
     .then(image => {
@@ -37,6 +43,7 @@ app.post('/foaming', function(req, res) {
 
 // post new notFoaming image to DB
 app.post('/notFoaming', function(req, res) {
+  console.log('receiving NOT foaming POST request')
   const {url, time, index, tag} = req.body;
   postImage(url, time, tag, index)
     .then(image => {
@@ -51,6 +58,7 @@ app.post('/notFoaming', function(req, res) {
 app.get('/foaming', function(req, res) {
   getImages('foaming')
     .then(imageList => {
+      console.log('server side list of foaming images: ', imageList)
       res.status(200).send(imageList);
     })
     .catch(err => {
@@ -62,6 +70,7 @@ app.get('/foaming', function(req, res) {
 app.get('/notFoaming', function(req, res) {
   getImages('notFoaming')
     .then(imageList => {
+      console.log('server side list of not foaming images: ', imageList)
       res.status(200).send(imageList);
     })
     .catch(err => {
@@ -69,7 +78,7 @@ app.get('/notFoaming', function(req, res) {
     })
 })
 
-const port = 3000;
+const port = 3001;
 
 app.listen(port, function() {
   console.log(`listening on port ${port}`);
